Updated function to main
the readragged function is now in the main. and read ragged .c is deleted
