+++ Code/main.c
@@ -1,43 +1,81 @@
-#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 
-struct character {
-  uint8_t level;
-  uint8_t spells[]; // spell level //amount of spells
+struct spellcard {
+  char *spel;
+};
+// structs for the player with pointers
+struct playerstats {
+  int *spellslots[9];
+  int *level;
+  int *spelamount;
+  struct spellcard *amount;
 };
 
-// make form json file structs
-// read ragged array
-void readargv(int argc, char *argv[], struct character *p);
+void readragged(int argc, char *argv[], struct playerstats); // read ragged array function
+
 
 int main(int argc, char *argv[]) {
-  struct character player;
 
-  readargv(argc, argv, &player);
-  printf("Done\n");
-  printf("%d\n", player.level);
+  
+  struct playerstats player1;
+  readragged(argc, &*argv, player1); // reading ragged array
 
-  return 0;
+
 }
 
-void readargv(int argc, char *argv[], struct character *p) {
+
+
+
+
+
+
+
+
+
+//----------------------------------------
+// read function
+void readragged(int argc, char *argv[], struct playerstats player1) {
 
   for (int i = 0; i < argc; i++) {
-    printf("%s/n", argv[i]);
 
-    if (strncmp(argv[i], "-s", 2)) {
-      p->spells[0] = atoi(argv[i + 1]);
-      p->spells[1] = atoi(argv[i + 2]);
-      p->spells[2] = atoi(argv[i + 3]);
-      p->spells[3] = atoi(argv[i + 4]);
-      p->spells[4] = atoi(argv[i + 5]);
-      p->spells[5] = atoi(argv[i + 6]);
-      p->spells[6] = atoi(argv[i + 7]);
-      p->spells[7] = atoi(argv[i + 8]);
-      p->spells[8] = atoi(argv[i + 9]);
-      p->spells[9] = atoi(argv[i + 10]);
+    if (argv[i] != NULL) {
+      if (strcmp(argv[i], "-s") == 0) {
+        for (int j = 1; j <= 9; j++) {
+          if (strcmp(argv[j + i], "-l") == 0 ||
+              (strcmp(argv[j + i], "-h") == 0 || argv[j + i] == NULL)) {
+            break;
+          }
+          player1.spellslots[j] = calloc(1, sizeof(int));
+          *player1.spellslots[j] = atoi(argv[j + i]);
+          printf("%d\n", *player1.spellslots[j]);
+        }
+      } else if (strcmp(argv[i], "-l") == 0) {
+        int count = 0;
+        for (int j = 1; j < 20; j++) {
+          if (strcmp(argv[j + i], "-h") == 0 || argv[j + i] == NULL) {
+            break;
+          }
+          if (j == 1) {
+            player1.level = calloc(1, sizeof(int));
+            *player1.level = atoi(argv[j + i]);
+            printf("%d\n", *player1.level);
+          } else
+            count++;
+        }
+        player1.spelamount = calloc(1, sizeof(int));
+        *player1.spelamount = count;
+        player1.amount = (struct spellcard *)calloc(*player1.spelamount, sizeof(struct spellcard));
+        for (int j = 2; j <2+ *player1.spelamount; j++) {
+          player1.amount[j].spel = (char *)calloc(strlen(argv[j + i]) + 1, sizeof(char));
+          strcpy(player1.amount[j].spel, argv[j + i]);
+          printf("%s\n", player1.amount[j].spel);
+        }
+      }
+     else if (strcmp(argv[i], "-h") == 0) {
+      printf("test3\n");
+     }
     }
   }
-};
+}
